Resolved conflicts in element adapters after source_sink → node rename:
- grid.py, inverter.py, load.py, solar.py: Updated element_type to 'node'
- node.py: Fixed variable naming consistency
- haeo.py extractor: Updated return type to use int timestamps

Updated battery test data for sensor field changes.
Added missing translations for battery sections and node switches.

+++ custom_components/haeo/data/loader/extractors/haeo.py
@@ -109,7 +109,7 @@ class Parser:
     @staticmethod
     def extract(
         state: HaeoForecastState,
-    ) -> tuple[Sequence[tuple[float, float]], str, SensorDeviceClass | None]:
+    ) -> tuple[Sequence[tuple[int, float]], str, SensorDeviceClass | None]:
         """Extract forecast data from HAEO forecast format.
 
         Expects list format: list of {"time": ..., "value": ...} dicts.
